refactor: split sim image editor into package
Break the 3078-line sd_create_simulated_image.py QMainWindow into a
cli/sim_editor/ package with one module per schema block: global fields,
noise, stray light, background stars, body/ring/star tabs, tab management,
render/display, and scene I/O, each a mixin over a shared SimEditorBase.
sd_create_simulated_image.py becomes the thin console-script dispatch that
re-exports the window and main(). Behavior-preserving; every module is well
under the 1000-line ceiling, so later phases add a control tab as a new
module rather than editing a monolith.
